refactor: apply folder structure
